chore: update struct

+++ src/structs/JBGovernanceNFTMint.sol
@@ -2,6 +2,6 @@
 pragma solidity ^0.8.17;
 
 struct JBGovernanceNFTMint {
-    uint200 stakeAmount;
+    uint256 stakeAmount;
     address beneficiary;
 }
